Wrap up program
Here I added the call to the game function, as well as some console.log
to see what the computer chose to make sure the program is working
correctly. I also added a loop for calling the game function, this way
the player can keep playing until they choose to type "no".