Finish styles and start code
This commit finishes the style for the page, as well as the overall
structure. It's not pretty, but will have to do for now.

I also started coding a little to get the hang of manipulating DOM.
Next I will add event listeners to the buttons to begin the game. The
idea is to display both the player's and computer's choice in the page
and keep track of the score.